Renaming test files to make more sense
